Changed methods in Account class so that they only do the function-part, added new methods in UI class that do the visual parts.

+++ OpenDefendersATM/Loan.cs
@@ -8,10 +8,10 @@ namespace OpenDefendersATM
 {
     internal class Loan
     {
-        public float MaxLoan { get; set; }
+        public decimal MaxLoan { get; set; }
         public int LoanID { get; set; }
-        public float Amount { get; set; }
-        public float InterestRate { get; set; }
+        public decimal Amount { get; set; }
+        public decimal InterestRate { get; set; }
         public int customerID { get; set; }
 
         public void CalculatePayment()
@@ -23,11 +23,11 @@ namespace OpenDefendersATM
         public void CheckLoanLimit(Customer customer)
         {
             Console.WriteLine("Ange det lönebelopp du vill låna:");
-            float loanLimit = customer.TotalBalance();
+            decimal loanLimit = customer.TotalBalance();
             MaxLoan = loanLimit * 5;
 
-            float userInput;
-            while (!float.TryParse(Console.ReadLine(), out userInput) || userInput <= 0 || userInput > MaxLoan)
+            decimal userInput;
+            while (!decimal.TryParse(Console.ReadLine(), out userInput) || userInput <= 0 || userInput > MaxLoan)
             {
                 Console.WriteLine("Du kan inte låna ett negativt belopp. Vänligen ange din önskad lånesumma." +
                     "Du får låna max 5 gånger ditt totala saldo");
